pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "univ-info-cursor"
version = "0.1.0"
description = "累積オープンキャンパス情報の取得・差分通知"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "pyyaml>=6.0",
  "typer>=0.12",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6"]

[project.scripts]
univ-oc-cursor = "univ_oc.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
